requirements.txt
fastapi
motor
uvicorn
pydantic-settings
passlib
pymongo
pyjwt
cloudinary